+++ app/controls/ContextMenu.svelte
@@ -3,6 +3,7 @@
     import type { Operand } from "../messages/Operand";
     import type { RevHeader } from "../messages/RevHeader";
     import type { StoreRef } from "../messages/StoreRef";
+    import { altKeyPressed } from "../stores";
     import RevisionMutator from "../mutators/RevisionMutator";
     import ChangeMutator from "../mutators/ChangeMutator";
     import RefMutator from "../mutators/RefMutator";
@@ -21,13 +22,14 @@
         return [];
     }
 
-    function onClick(action: string) {
+    function onClick(action: string, event: MouseEvent) {
+        let ignoreImmutable = event.altKey || $altKeyPressed;
         if (operand.type === "Revision" || operand.type === "Revisions") {
-            new RevisionMutator(getRevisionHeaders()).handle(action);
+            new RevisionMutator(getRevisionHeaders(), ignoreImmutable).handle(action);
         } else if (operand.type === "Change") {
-            new ChangeMutator(operand.headers, operand.path, operand.hunk).handle(action);
+            new ChangeMutator(operand.headers, operand.path, operand.hunk, ignoreImmutable).handle(action);
         } else if (operand.type === "Ref") {
-            new RefMutator(operand.ref).handle(action);
+            new RefMutator(operand.ref, ignoreImmutable).handle(action);
         }
         onClose();
     }
@@ -42,9 +44,9 @@
         }
     }
 
-    function isRevisionEnabled(headers: RevHeader[]) {
+    function isRevisionEnabled(headers: RevHeader[], altHeld: boolean = false) {
         const isSingleton = headers.length == 1;
-        const anyImmutable = headers.some((h) => h.is_immutable);
+        const anyImmutable = !altHeld && headers.some((h) => h.is_immutable);
         const hasSingleParent = headers[headers.length - 1]?.parent_ids.length == 1;
 
         return {
@@ -60,8 +62,8 @@
         };
     }
 
-    function isChangeEnabled(headers: RevHeader[]) {
-        const anyImmutable = headers.some((h) => h.is_immutable);
+    function isChangeEnabled(headers: RevHeader[], altHeld: boolean = false) {
+        const anyImmutable = !altHeld && headers.some((h) => h.is_immutable);
         const hasSingleParent = headers[headers.length - 1]?.parent_ids.length == 1;
 
         return {
@@ -90,8 +92,10 @@
     }
 
     $: revisionEnabled =
-        operand.type === "Revision" || operand.type === "Revisions" ? isRevisionEnabled(getRevisionHeaders()) : null;
-    $: changeEnabled = operand.type === "Change" ? isChangeEnabled(operand.headers) : null;
+        operand.type === "Revision" || operand.type === "Revisions" ? isRevisionEnabled($selectionHeaders) : null;
+            ? isRevisionEnabled($selectionHeaders, $altKeyPressed)
+            : null;
+    $: changeEnabled = operand.type === "Change" ? isChangeEnabled(operand.headers, $altKeyPressed) : null;
     $: refEnabled = operand.type === "Ref" ? isRefEnabled(operand.ref) : null;
 
     // clamp to viewport
@@ -121,35 +125,37 @@
     on:keydown={onKeyDown}
     bind:this={menuElement}>
     {#if (operand.type === "Revision" || operand.type === "Revisions") && revisionEnabled}
-        <button disabled={!revisionEnabled.new_child} on:click={() => onClick("new_child")}>New child</button>
-        <button disabled={!revisionEnabled.new_parent} on:click={() => onClick("new_parent")}
+        <button disabled={!revisionEnabled.new_child} on:click={(e) => onClick("new_child", e)}>New child</button>
+        <button disabled={!revisionEnabled.new_parent} on:click={(e) => onClick("new_parent", e)}
             >New inserted parent</button>
         <hr />
-        <button disabled={!revisionEnabled.edit} on:click={() => onClick("edit")}>Edit as working copy</button>
-        <button disabled={!revisionEnabled.revert} on:click={() => onClick("revert")}
+        <button disabled={!revisionEnabled.edit} on:click={(e) => onClick("edit", e)}>Edit as working copy</button>
+        <button disabled={!revisionEnabled.revert} on:click={(e) => onClick("revert", e)}
             >Revert into working copy</button>
-        <button disabled={!revisionEnabled.duplicate} on:click={() => onClick("duplicate")}>Duplicate</button>
-        <button disabled={!revisionEnabled.abandon} on:click={() => onClick("abandon")}>Abandon</button>
+        <button disabled={!revisionEnabled.duplicate} on:click={(e) => onClick("duplicate", e)}>Duplicate</button>
+        <button disabled={!revisionEnabled.abandon} on:click={(e) => onClick("abandon", e)}>Abandon</button>
         <hr />
-        <button disabled={!revisionEnabled.squash} on:click={() => onClick("squash")}>Squash into parent</button>
-        <button disabled={!revisionEnabled.restore} on:click={() => onClick("restore")}>Restore from parent</button>
+        <button disabled={!revisionEnabled.squash} on:click={(e) => onClick("squash", e)}>Squash into parent</button>
+        <button disabled={!revisionEnabled.restore} on:click={(e) => onClick("restore", e)}>Restore from parent</button>
         <hr />
-        <button disabled={!revisionEnabled.bookmark} on:click={() => onClick("bookmark")}>Create bookmark...</button>
+        <button disabled={!revisionEnabled.bookmark} on:click={(e) => onClick("bookmark", e)}
+            >Create bookmark...</button>
     {:else if operand.type === "Change" && changeEnabled}
-        <button disabled={!changeEnabled.squash} on:click={() => onClick("squash")}>Squash into parent</button>
-        <button disabled={!changeEnabled.restore} on:click={() => onClick("restore")}>Restore from parent</button>
+        <button disabled={!changeEnabled.squash} on:click={(e) => onClick("squash", e)}>Squash into parent</button>
+        <button disabled={!changeEnabled.restore} on:click={(e) => onClick("restore", e)}>Restore from parent</button>
     {:else if operand.type === "Ref" && refEnabled}
-        <button disabled={!refEnabled.track} on:click={() => onClick("track")}>Track</button>
-        <button disabled={!refEnabled.untrack} on:click={() => onClick("untrack")}>Untrack</button>
+        <button disabled={!refEnabled.track} on:click={(e) => onClick("track", e)}>Track</button>
+        <button disabled={!refEnabled.untrack} on:click={(e) => onClick("untrack", e)}>Untrack</button>
         <hr />
-        <button disabled={!refEnabled.push_all} on:click={() => onClick("push-all")}>Push</button>
-        <button disabled={!refEnabled.push_single} on:click={() => onClick("push-single")}>Push to remote...</button>
-        <button disabled={!refEnabled.fetch_all} on:click={() => onClick("fetch-all")}>Fetch</button>
-        <button disabled={!refEnabled.fetch_single} on:click={() => onClick("fetch-single")}
+        <button disabled={!refEnabled.push_all} on:click={(e) => onClick("push-all", e)}>Push</button>
+        <button disabled={!refEnabled.push_single} on:click={(e) => onClick("push-single", e)}
+            >Push to remote...</button>
+        <button disabled={!refEnabled.fetch_all} on:click={(e) => onClick("fetch-all", e)}>Fetch</button>
+        <button disabled={!refEnabled.fetch_single} on:click={(e) => onClick("fetch-single", e)}
             >Fetch from remote...</button>
         <hr />
-        <button disabled={!refEnabled.rename} on:click={() => onClick("rename")}>Rename...</button>
-        <button disabled={!refEnabled.delete} on:click={() => onClick("delete")}>Delete</button>
+        <button disabled={!refEnabled.rename} on:click={(e) => onClick("rename", e)}>Rename...</button>
+        <button disabled={!refEnabled.delete} on:click={(e) => onClick("delete", e)}>Delete</button>
     {/if}
 </div>